Add a PR template for improved understanding of such a PR
Without a structured or pattern pull requests tend to be obscure
and vary in the amount of details they provide. Making it difficult
to know what is being addressed and whether those features are required
or valid.

But with this template, all PRs will follow the same pattern and will
provide details such as:

* What the PR does
* Description of the task that the PR will address
* Associated Pivotal Tracker stories
* Relevant background context as required in the PR
* Screenshots

These details will make the PRs concise, easy to follow and understand