Fix a type in index.js causing a 500 server error
When I tried accessing the 'Food List' nav link on the /admin
section of the web site, it throws a 500 code server error.
Upon investigating, I noticed I have a typographical error when
returning that route's response.
I wrote resp, instead of responseObj. That was what caused the error
which this commit will fixed