Add a scrollbar to the order summary display
As it is, when customers add items to their orders,
the order summary div that displays the orders - food, pric, quantity-
grows as the list grows.
When many items are added, some order items will go off the screen and
there's no scrolling mechanism to bring them back into view.

This commit addresses that by adding a scrollbar to the div, such that
when the many items are added and the set height of the div
is reached, the scroll bar automatically becomes active, allowing
vertical scrolling