Add babel to compile all js files from ES6 to ES5
All the javascript files in the project were/are being in written in
ES6 implementation. Due to the fact that javascript versions ship out
faster than browsers can keep track, its possible to write a js code
and a browser will be t a loss as to what to do with it.

That's where the Babel comes in. It will compile the js files from
ES6 and compile it to ES5, hence giving a whole lot of browsers the
ability to safely and properly execute the js files.