Fix status update button on admin panel
Commit fixes the issue with the the status update button
not showing up. The button was created in userorder.js.

Fixed it by declaring and hooking it up with it's accompanying table
cell imediately. That solved it as declaring and attaching the two
at different places broke in es5 after babel compilation