Modify project structure to use models and controllers
Will add models and controllers to the project setup. This is to
facilitate modularity of the project.
Initially, all actions were being done on the index.js file.
But now, the routes call the conrollers, which in turn calls the
models, passing in any required parameter and getting the available
responses