Add PostgreSQL to the local copy of project
Tried experimenting with adding PostgreSQL to the site,
as per the project specs. Have been able to implement it
(a users table), with it's accompanying controller and model. Next task
is to implement same in the project itself.
NOTE:
This branch is not to be merged with the develop or any other branch